+++ proofs/clock.php
@@ -116,4 +116,29 @@ return static function() {
             ),
         ),
     );
+
+    yield proof(
+        'Clock::ofFormat()->at()',
+        given(PointInTime::any()),
+        static function($assert, $point) {
+            $parsed = Clock::live()
+                ->ofFormat(Format::iso8601())
+                ->at($point->format(Format::iso8601()))
+                ->match(
+                    static fn($point) => $point->format(Format::iso8601()),
+                    static fn() => null,
+                );
+
+            $assert->not()->null($parsed);
+            $assert->same(
+                Clock::live()
+                    ->at($point->format(Format::iso8601()), Format::iso8601())
+                    ->match(
+                        static fn($point) => $point->format(Format::iso8601()),
+                        static fn() => null,
+                    ),
+                $parsed,
+            );
+        },
+    );
 };